platform/my-vuln-app/server/requirements.txt
alembic==1.7.4
aniso8601==9.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
click==8.0.3
distlib==0.3.3
filelock==3.3.1
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
greenlet==1.1.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.0
Mako==1.1.5
MarkupSafe==2.0.1
packaging==21.0
pipenv==2021.5.29
platformdirs==2.4.0
pluggy==1.0.0
py==1.10.0
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
python-dotenv==0.19.1
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.25
toml==0.10.2
virtualenv==20.8.1
virtualenv-clone==0.5.7
Werkzeug==2.0.2
